Moved 5 funcs to `get_clan_cats.py`: get_alive_clan_queens, find_alive_cats_with_rank, get_living_cat_count, get_living_clan_cat_count, get_cats_same_age, get_free_possible_mates
