+++ scripts/clan.py
@@ -44,8 +44,8 @@ from scripts.housekeeping.version import get_version_info, SAVE_VERSION_NUMBER
 from scripts.utility import (
     get_current_season,
     clan_symbol_sprite,
-    get_living_clan_cat_count,
 )  # pylint: disable=redefined-builtin
+from scripts.clan_package.get_clan_cats import get_living_clan_cat_count
 
 
 class Clan:
