+++ scripts/events_module/event_filters.py
@@ -6,9 +6,9 @@ from scripts.cat_relations.enums import RelType
 from scripts.cat.enums import CatRank, CatAge
 from scripts.special_dates import get_special_date, contains_special_date_tag
 from scripts.utility import (
-    find_alive_cats_with_rank,
     filter_relationship_type,
 )
+from scripts.clan_package.get_clan_cats import find_alive_cats_with_rank
 from scripts.game_structure import game
