+++ scripts/events_module/ongoing/disaster_events.py
@@ -5,7 +5,7 @@ from scripts.cat.enums import CatRank
 from scripts.event_class import Single_Event
 from scripts.events_module.generate_events import GenerateEvents
 from scripts.game_structure import game
-from scripts.utility import find_alive_cats_with_rank
+from scripts.clan_package.get_clan_cats import find_alive_cats_with_rank
 
 
 # ---------------------------------------------------------------------------- #
