+++ scripts/events_module/short/condition_events.py
@@ -32,9 +32,9 @@ from scripts.game_structure import game
 from scripts.game_structure.localization import load_lang_resource
 from scripts.utility import (
     event_text_adjust,
-    find_alive_cats_with_rank,
     get_leader_life_notice,
 )
+from scripts.clan_package.get_clan_cats import find_alive_cats_with_rank
 
 
 # ---------------------------------------------------------------------------- #
