+++ scripts/screens/AllegiancesScreen.py
@@ -8,11 +8,11 @@ from scripts.game_structure.screen_settings import MANAGER
 from scripts.utility import (
     get_text_box_theme,
     ui_scale,
-    get_alive_clan_queens,
     ui_scale_offset,
     adjust_list_text,
     event_text_adjust,
 )
+from ..clan_package.get_clan_cats import get_alive_clan_queens
 from .Screens import Screens
 from ..cat.enums import CatRank
 from ..game_structure.ui_elements import UIModifiedScrollingContainer
