+++ tests/test_events.py
@@ -14,7 +14,7 @@ from scripts.events_module.short.short_event_generation import (
     filter_events,
 )
 from scripts.game_structure import game
-from scripts.utility import get_living_cat_count
+from scripts.clan_package.get_clan_cats import get_living_cat_count
 
 
 class TestEvents(unittest.TestCase):
@@ -93,4 +93,5 @@ class TestEvents(unittest.TestCase):
             for _ in range(500):
                 events.one_moon()
                 if not _ % 100:
+                    # todo: why use this function? this also counts COTC, which is useless for telling if the clan is dead
                     print(f"CATS ALIVE: {get_living_cat_count(Cat)}")
