+++ tests/test_freshkill.py
@@ -11,7 +11,7 @@ from scripts.cat.cats import Cat
 from scripts.cat.skills import Skill, SkillPath
 from scripts.clan import Clan
 from scripts.clan_resources.freshkill import FreshkillPile
-from scripts.utility import get_alive_clan_queens
+from scripts.clan_package.get_clan_cats import get_alive_clan_queens
 
 
 class FreshkillPileTest(unittest.TestCase):
